Agrego funcion GET_DATA a los templates. Agrego contadores de update y insert total. Agrego comprobacion de existencia en deletes genericos

+++ app/cfg/templates/general-delete-script-despliegue.sql
@@ -1,1 +1,4 @@
-	delete from {{ TABLA }} where {% for campo in CLAVES %} {{ campo }} = {{ VALORES[campo] }} {% if !loop.last %} and {% endif %}{% endfor %};
+	select count(1) into V_COUNT from {{ TABLA }} where {% for campo in CLAVES %} {{ campo }} = {{ VALORES[campo] }} {% if !loop.last %} and {% endif %}{% endfor %};
+	if V_COUNT > 0 then
+		delete from {{ TABLA }} where {% for campo in CLAVES %} {{ campo }} = {{ VALORES[campo] }} {% if !loop.last %} and {% endif %}{% endfor %};
+	end if;
